+++ src/fr/diginamic/utils/SalarieFactory.java
@@ -5,10 +5,7 @@ import fr.diginamic.entites.Salarie;
 public class SalarieFactory {
 	
 	public static Salarie createSalarie(String str) {
-		
 		String[] tabSalarie = str.split(";"); 	
-		
-		return new Salarie(tabSalarie[0], tabSalarie[1], Double.parseDouble(tabSalarie[2].replace(" ", "")));
+		return new Salarie(tabSalarie[0], tabSalarie[1], Float.parseFloat(tabSalarie[2].replace(" ", "")));
 	}
-
 }
